requirements.txt
PyAutoGUI
opencv-python
python-dotenv

#pip install opencv-python
#pip install -r requirements.txt